chore(frontend): rename LEI status filter util to kebab-case
